+++ splitwise-csharp/Interfaces/ISplitwiseUser.cs
@@ -1,0 +1,13 @@
+﻿using SplitwiseCSharp.Models;
+
+namespace SplitwiseCSharp.Interfaces;
+
+public interface ISplitwiseUser
+{
+    int Id { get; set; }
+    string? FirstName { get; set; }
+    string? LastName { get; set; }
+    string? Email { get; set; }
+    string? RegistrationStatus { get; set; }
+    SplitwisePicture? Picture { get; set; }
+}
