+++ splitwise-csharp/Models/SplitwiseUser.cs
@@ -1,6 +1,11 @@
-﻿namespace SplitwiseCSharp.Models;
+﻿using SplitwiseCSharp.Interfaces;
 
-public class SplitwiseUser
+namespace SplitwiseCSharp.Models;
+
+/// <summary>
+/// Base information available about all users through ID lookup.
+/// </summary>
+public class SplitwiseUser : ISplitwiseUser
 {
     public int Id { get; set; }
     public string? FirstName { get; set; }
@@ -8,19 +13,6 @@ public class SplitwiseUser
     public string? Email { get; set; }
     public string? RegistrationStatus { get; set; }
     public SplitwisePicture? Picture { get; set; }
-    public DateTime NotificationsRead { get; set; }
-    public int NotificationsCount { get; set; }
-    public SplitwiseNotificationSettings? Notifications { get; set; }
-    public string? DefaultCurrency { get; set; }
-    public string? Locale { get; set; }
-    public string CountryCode { get; set; }
-    public string DateFormat { get; set; }
-    public DateTime ForceRefreshAt { get; set; }
-    public bool CustomPicture { get; set; }
-    public int DefaultGroupId { get; set; }
-
 }
 
-public class SplitwiseNotificationSettings
-{
-}
+
