--- splitwise-csharp/Requests/GetAccessTokenRequest.cs
@@ -1,7 +1,0 @@
-namespace SplitwiseCSharp.Requests;
-public class GetAccessTokenRequest
-{
-    public required string grant_type { get; set; }
-    public required string client_id { get; set; }
-    public required string client_secret { get; set; }
-}
