--- splitwise-csharp/Requests/GetAuthCodeRequest.cs
@@ -1,8 +1,0 @@
-namespace SplitwiseCSharp.Requests;
-public class GetAuthCodeRequest
-{
-    public required string grant_type { get; set; }
-    public required string code { get; set; }
-    public required string client_id { get; set; }
-    public required string client_secret { get; set; }
-}
