+++ splitwise-csharp/Responses/GetUserByIdResponse.cs
@@ -1,0 +1,12 @@
+﻿using SplitwiseCSharp.Models;
+
+
+namespace SplitwiseCSharp.Responses;
+
+/// <summary>
+/// A wrapper for a response from the GetUser endpoint.
+/// </summary>
+public class GetUserByIdResponse : SplitwiseUser
+{
+
+}
