+++ splitwise-csharp/Responses/SplitwiseTokenResponse.cs
@@ -2,7 +2,10 @@
 
 namespace SplitwiseCSharp.Responses;
 
-public class SplitwiseTokenResponse
+/// <summary>
+/// A wrapper class for the response from generating an auth token from Splitwise.
+/// </summary>
+internal class SplitwiseTokenResponse
 {
     [JsonPropertyName("access_token")]
     public string? AccessToken { get; set; }
